backend/requirements.txt
celery
django
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary
pylint
pylint-django
requests
requests-toolbelt
redis